requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
groq==0.4.1
pydantic==2.4.2
python-multipart==0.0.6
httpx>=0.24.1
typing-extensions>=4.5.0 